setup databases on startup
